# api/swagger.4.0.0.json
{
    "swagger": "2.0",
    "info": {
        "title": "Intent",
        "description": "This is Swagger UI environment generated for the TM Forum Intent Management API specification",
        "version": "4.0.0"
    },
    "host": "serverRoot",
    "basePath": "/tmf-api/intent/v4/",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json;charset=utf-8"
    ],
    "produces": [
        "application/json;charset=utf-8"
    ],
...
    "paths": {
...
        "/intent/{intentId}/intentReport": {
            "get": {
                "operationId": "listIntentReport",
                "summary": "List or find IntentReport objects",
                "description": "This operation list or find IntentReport entities",
                "tags": [
                    "intentReport"
                ],
                "parameters": [
                    {
                        "name": "intentId",
                        "description": "Identifier of the parent Intent entity",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "fields",
                        "description": "Comma-separated properties to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "offset",
                        "description": "Requested index for start of resources to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "description": "Requested number of resources to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "X-Result-Count": {
                                "description": "Actual number of items returned in the response body",
                                "type": "integer"
                            },
                            "X-Total-Count": {
                                "description": "Total number of items matching criteria",
                                "type": "integer"
                            }
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/IntentReport"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "operationId": "createIntentReport",
                "summary": "Creates a IntentReport",
                "description": "This operation creates a IntentReport entity.",
                "tags": [
                    "intentReport"
                ],
                "parameters": [
                    {
                        "name": "intentId",
                        "description": "Identifier of the parent Intent entity",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "intentReport",
                        "description": "The IntentReport to be created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/IntentReport_Create"
                        },
                        "in": "body"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/IntentReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/intent/{intentId}/intentReport/{id}": {
            "get": {
                "operationId": "retrieveIntentReport",
                "summary": "Retrieves a IntentReport by ID",
                "description": "This operation retrieves a IntentReport entity. Attribute selection is enabled for all first level attributes.",
                "tags": [
                    "intentReport"
                ],
                "parameters": [
                    {
                        "name": "intentId",
                        "description": "Identifier of the parent Intent entity",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "id",
                        "description": "Identifier of the IntentReport",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "fields",
                        "description": "Comma-separated properties to provide in response",
                        "required": false,
                        "type": "string",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/IntentReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "patch": {
                "operationId": "patchIntentReport",
                "summary": "Updates partially a IntentReport",
                "description": "This operation updates partially a IntentReport entity.",
                "tags": [
                    "intentReport"
                ],
                "parameters": [
                    {
                        "name": "intentId",
                        "description": "Identifier of the parent Intent entity",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "id",
                        "description": "Identifier of the IntentReport",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "intentReport",
                        "description": "The IntentReport to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/IntentReport_Update"
                        },
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/IntentReport"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteIntentReport",
                "summary": "Deletes a IntentReport",
                "description": "This operation deletes a IntentReport entity.",
                "tags": [
                    "intentReport"
                ],
                "parameters": [
                    {
                        "name": "intentId",
                        "description": "Identifier of the parent Intent entity",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "id",
                        "description": "Identifier of the IntentReport",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "IntentReport": {
            "type": "object",
            "description": "IntentReport is the mechanism to report back to Intent owner on an Intents status",
            "properties": {
                "id": {
                    "type": "string",
                    "description": "unique identifier"
                },
                "href": {
                    "type": "string",
                    "format": "uri",
                    "description": "Hyperlink reference"
                },
                "creationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Date and time of the creation of this intent report"
                },
                "description": {
                    "type": "string",
                    "description": "The description of the intent report."
                },
                "name": {
                    "type": "string",
                    "description": "The name of the intent report."
                },
                "expression": {
                    "$ref": "#/definitions/Expression",
                    "description": "An Intent Report contains a set of expectations expressed  (Expression) in statements using a particular espressionLanguage. This attribute captures the intent report expressed by the TM Forum Intent Ontology (TIO) as is stored as a string. The Expression is the formal statement of the intent and contains all the information that describes the complete intent."
                },
                "intent": {
                    "$ref": "#/definitions/Intent"
                },
                "validFor": {
                    "$ref": "#/definitions/TimePeriod",
                    "description": "The period for which this intent report is valid"
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
        "IntentReport_Create": {
            "type": "object",
            "description": "IntentReport is the mechanism to report back to Intent owner on an Intents status\nSkipped properties: id,href",
            "required": [
                "name",
                "expression",
                "intent"
            ],
            "properties": {
                "creationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Date and time of the creation of this intent report"
                },
                "description": {
                    "type": "string",
                    "description": "The description of the intent report."
                },
                "name": {
                    "type": "string",
                    "description": "The name of the intent report."
                },
                "expression": {
                    "$ref": "#/definitions/Expression",
                    "description": "An Intent Report contains a set of expectations expressed  (Expression) in statements using a particular espressionLanguage. This attribute captures the intent report expressed by the TM Forum Intent Ontology (TIO) as is stored as a string. The Expression is the formal statement of the intent and contains all the information that describes the complete intent."
                },
                "intent": {
                    "$ref": "#/definitions/Intent"
                },
                "validFor": {
                    "$ref": "#/definitions/TimePeriod",
                    "description": "The period for which this intent report is valid"
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
        "IntentReport_Update": {
            "type": "object",
            "description": "IntentReport is the mechanism to report back to Intent owner on an Intents status\nSkipped properties: id,href",
            "properties": {
                "creationDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Date and time of the creation of this intent report"
                },
                "description": {
                    "type": "string",
                    "description": "The description of the intent report."
                },
                "name": {
                    "type": "string",
                    "description": "The name of the intent report."
                },
                "expression": {
                    "$ref": "#/definitions/Expression",
                    "description": "An Intent Report contains a set of expectations expressed  (Expression) in statements using a particular espressionLanguage. This attribute captures the intent report expressed by the TM Forum Intent Ontology (TIO) as is stored as a string. The Expression is the formal statement of the intent and contains all the information that describes the complete intent."
                },
                "intent": {
                    "$ref": "#/definitions/Intent"
                },
                "validFor": {
                    "$ref": "#/definitions/TimePeriod",
                    "description": "The period for which this intent report is valid"
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
...
        "Error": {
            "description": "Used when an API throws an Error, typically with a HTTP error response-code (3xx, 4xx, 5xx)",
            "type": "object",
            "required": [
                "code",
                "reason"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "description": "Application relevant detail, defined in the API or a common list."
                },
                "reason": {
                    "type": "string",
                    "description": "Explanation of the reason for the error which can be shown to a client user."
                },
                "message": {
                    "type": "string",
                    "description": "More details and corrective actions related to the error which can be shown to a client user."
                },
                "status": {
                    "type": "string",
                    "description": "HTTP Error code extension"
                },
                "referenceError": {
                    "type": "string",
                    "format": "uri",
                    "description": "URI of documentation describing the error."
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class."
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class entity name."
                }
            }
        }
    }
}